third_party/WebKit/LayoutTests/animations/animation-finish-promise-gc.html

<!DOCTYPE html>
<meta charset=utf-8>
<title>Garbage collect an animation with a finish promise</title>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<script src="../resources/gc.js"></script>
<body>
<script>
async_test(t => {
  assert_own_property(window, "GCController", "gc controller needed");
  setInterval(gc, 0);

  var anim = document.body.animate(null, 1);
  var promise = anim.finished;

  promise.then(() => {
    t.done();
  });
}, 'Animation finish promise resolved if animation is garbage collected');
</script>
</body>
